+++ src/services/item/plugins/file/service.ts
@@ -1,6 +1,7 @@
-import fs from 'fs';
 import path from 'path';
+import { PassThrough, Readable } from 'stream';
 
+import { MultipartFields, MultipartFile } from '@fastify/multipart';
 import { FastifyReply } from 'fastify';
 
 import {
@@ -16,8 +17,8 @@ import { UnauthorizedMember } from '../../../../utils/errors';
 import { Repositories } from '../../../../utils/repositories';
 import { validatePermission } from '../../../authorization';
 import FileService from '../../../file/service';
+import { UploadEmptyFileError } from '../../../file/utils/errors';
 import { Actor, Member } from '../../../member/entities/member';
-import { UploadedFile } from '../../../thumbnail/types';
 import { randomHexOf4 } from '../../../utils';
 import { Item } from '../../entities/Item';
 import ItemService from '../../service';
@@ -59,11 +60,7 @@ class FileItemService {
 
   // check the user has enough storage to create a new item given its size
   // get the complete storage
-  async checkRemainingStorage(actor: Member, repositories: Repositories, size?: number) {
-    if (!size) {
-      return;
-    }
-
+  async checkRemainingStorage(actor: Member, repositories: Repositories, size: number = 0) {
     const { id: memberId } = actor;
 
     const currentStorage = await repositories.memberRepository.getMemberStorage(
@@ -72,74 +69,71 @@ class FileItemService {
     );
 
     if (currentStorage + size > this.options.maxMemberStorage) {
-      throw new StorageExceeded();
+      throw new StorageExceeded(currentStorage + size);
     }
   }
 
-  async upload(actor: Actor, repositories: Repositories, files: UploadedFile[], parentId: string) {
-    if (!actor) {
-      throw new UnauthorizedMember(actor);
-    }
-
-    // check rights
-    if (parentId) {
-      const item = await repositories.itemRepository.get(parentId);
-      await validatePermission(repositories, PermissionLevel.Write, actor, item);
-    }
-
-    const promises: Promise<{
-      filepath: string;
-      filename: string;
-      size: number;
-      mimetype: string;
-      uploaded: UploadedFile;
-    }>[] = [];
-    for (const fileObject of files) {
-      const { filename, mimetype, fields, filepath: tmpPath } = fileObject;
-      const file = fs.createReadStream(tmpPath);
-      const { size } = fs.statSync(tmpPath);
-      const filepath = this.buildFilePath(); // parentId, filename
-
-      // compute body data from file's fields
-      if (fields) {
-        const fileBody = Object.fromEntries(
-          Object.keys(fields).map((key) => [
-            key,
-            (fields[key] as unknown as { value: string })?.value,
-          ]),
-        );
-      }
-      // check member storage limit
-      await this.checkRemainingStorage(actor, repositories, size);
-
-      promises.push(
-        this.fileService
-          .upload(actor, {
-            file,
-            filepath,
-            mimetype,
-            size,
-          })
-          .then(() => {
-            return { filepath, filename, size, mimetype, uploaded: fileObject };
-          })
-          .catch((e) => {
-            throw e;
-          }),
+  async upload(
+    actor,
+    repositories,
+    {
+      description,
+      parentId,
+      filename,
+      mimetype,
+      fields,
+      stream,
+    }: {
+      description?: string;
+      parentId?: string;
+      filename;
+      mimetype;
+      fields?: MultipartFields;
+      stream: Readable;
+    },
+  ) {
+    const filepath = this.buildFilePath(); // parentId, filename
+    // compute body data from file's fields
+    if (fields) {
+      const fileBody = Object.fromEntries(
+        Object.keys(fields).map((key) => [
+          key,
+          (fields[key] as unknown as { value: string })?.value,
+        ]),
       );
     }
+    // check member storage limit
+    // BUG: this creates a big leak!!
+    // await this.checkRemainingStorage(actor, repositories);
 
-    // TODO: CHUNK TO AVOID FLOODING
-    // fallback?
-    const fileProperties = await Promise.all(promises);
+    // duplicate stream to use in thumbnails
+    const streamForThumbnails = new PassThrough();
+    if (MimeTypes.isImage(mimetype)) {
+      stream.pipe(streamForThumbnails);
+    }
 
-    const items: Item[] = [];
-    // postHook: create items from file properties
-    for (const properties of fileProperties) {
-      const { filename, filepath, mimetype, size, uploaded } = properties;
+    try {
+      await this.fileService.upload(actor, {
+        file: stream,
+        filepath,
+        mimetype,
+      });
+
+      const size = await this.fileService.getFileSize(actor, filepath);
+
+      // throw for empty files
+      if (!size) {
+        await this.fileService.delete(actor, filepath);
+        // Bug: empty stream does not trigger end event, which leads to hang in the for await
+        streamForThumbnails.emit('close');
+        throw new UploadEmptyFileError();
+      }
+
+      // create item from file properties
       const name = filename.substring(0, ORIGINAL_FILENAME_TRUNCATE_LIMIT);
       const item = {
         name,
+        description,
         type: this.fileService.type,
         extra: {
           [this.fileService.type]: {
@@ -158,17 +152,26 @@ class FileItemService {
         item,
         parentId,
       });
-      items.push(newItem);
 
       // add thumbnails if image
       // allow failures
       if (MimeTypes.isImage(mimetype)) {
-        await this.itemThumbnailService
-          .upload(actor, repositories, newItem.id, uploaded)
-          .catch((e) => console.error(e));
+        try {
+          await this.itemThumbnailService.upload(
+            actor,
+            repositories,
+            newItem.id,
+            streamForThumbnails,
+          );
+        } catch (e) {
+          console.error(e);
+        }
       }
+
+      return newItem;
+    } finally {
+      streamForThumbnails.emit('close');
     }
-    return items;
   }
 
   async download(
@@ -216,7 +219,8 @@ class FileItemService {
     };
 
     // check member storage limit
-    await this.checkRemainingStorage(actor, repositories, size);
+    // BUG: this creates a big leak!!
+    // await this.checkRemainingStorage(actor, repositories, size);
 
     // DON'T use task runner for copy file task: this would generate a new transaction
     // which is useless since the file copy task should not touch the DB at all