+++ src/services/item/plugins/published/constants.ts
@@ -1,9 +1,10 @@
 import { Context } from '@graasp/sdk';
 
-import { CLIENT_HOSTS, PROTOCOL } from '../../../../utils/config';
+import { CLIENT_HOSTS } from '../../../../utils/config';
 import { Item } from '../../entities/Item';
 
 export const buildPublishedItemLink = (item: Item) => {
   const library = CLIENT_HOSTS.find(({ name }) => name === Context.Library);
-  return `${PROTOCOL}://${library}/${item.id}`;
+  const target = new URL(item.id, library?.url);
+  return target.toString();
 };